@@ -7,7 +7,7 @@ namespace SingletonUsingMutex
     {
         static void Main()
         {
-            var mutex = new Mutex(false, "SingletonAppMutex", out bool mutexCreated);
+            Mutex mutex = new(false, "SingletonAppMutex", out bool mutexCreated);
             if (!mutexCreated)
             {
                 Console.WriteLine("You can only start one instance of the application.");
