@@ -8,8 +8,8 @@ namespace SynchronizatonSamples
         static void Main()
         {
             int numTasks = 20;
-            var state = new SharedState();
-            var tasks = new Task[numTasks];
+            SharedState state = new();
+            Task[] tasks = new Task[numTasks];
 
             for (int i = 0; i < numTasks; i++)
             {
